requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
email-validator==2.2.0
motor==3.5.1
python-dotenv==1.0.1
motor